requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
SQLAlchemy==2.0.38
supabase==2.13.0
psycopg2==2.9.10
python-dotenv==1.0.1
boto3==1.37.1
requests
aioitertools==0.12.0
aiohttp==3.11.13
pdfplumber==0.11.5
pdf2jpg==1.1
pyPDF2==3.0.1

